refactor: DispatcherServlet을 app 모듈에서 mvc 모듈로 이동
