+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServlet.java
@@ -1,6 +1,5 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
-import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.handler.HandlerAdapter;
 import com.interface21.webmvc.servlet.handler.HandlerExecutionHandlerAdapter;
 import com.interface21.webmvc.servlet.handler.HandlerExecutor;
